Rename Auth -> AuthCard
+ move debug stuff to App (todo remove later)
